requirements.txt
yowsup2
python-dotenv
protobuf==3.20.*